+++ build.gradle
@@ -34,7 +34,6 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation' // 유효성 검사
     implementation 'org.springframework.boot:spring-boot-starter-thymeleaf' // 템플릿 엔진
     implementation 'org.springframework.boot:spring-boot-starter-webflux' // WebFlux
-    developmentOnly 'org.springframework.boot:spring-boot-devtools' // 개발 도구
 
     // Security
     implementation 'org.springframework.boot:spring-boot-starter-security' // Spring Security
