--- src/main/java/page/clab/api/domain/memberManagement/member/application/port/in/ResendMemberPasswordUseCase.java
@@ -1,5 +1,0 @@
-package page.clab.api.domain.memberManagement.member.application.port.in;
-
-public interface ResendMemberPasswordUseCase {
-    String resendMemberPassword(String memberId);
-}
