--- src/main/java/page/clab/api/domain/memberManagement/member/application/service/ResetMemberPasswordVerificationService.java
@@ -1,37 +1,0 @@
-package page.clab.api.domain.memberManagement.member.application.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import page.clab.api.domain.memberManagement.member.application.port.in.VerifyResetMemberPasswordUseCase;
-import page.clab.api.domain.memberManagement.member.application.port.out.RegisterMemberPort;
-import page.clab.api.domain.memberManagement.member.application.port.out.RetrieveMemberPort;
-import page.clab.api.domain.memberManagement.member.domain.Member;
-import page.clab.api.global.common.verification.application.VerificationService;
-import page.clab.api.global.common.verification.domain.Verification;
-import page.clab.api.global.common.verification.dto.request.VerificationRequestDto;
-
-@Service
-@RequiredArgsConstructor
-public class ResetMemberPasswordVerificationService implements VerifyResetMemberPasswordUseCase {
-
-    private final RegisterMemberPort registerMemberPort;
-    private final RetrieveMemberPort retrieveMemberPort;
-    private final VerificationService verificationService;
-    private final PasswordEncoder passwordEncoder;
-
-    @Transactional
-    @Override
-    public String verifyResetMemberPassword(VerificationRequestDto requestDto) {
-        Member member = retrieveMemberPort.findByIdOrThrow(requestDto.getMemberId());
-        Verification verification = verificationService.validateVerificationCode(requestDto, member);
-        updateMemberPasswordWithVerificationCode(verification.getVerificationCode(), member);
-        return registerMemberPort.save(member).getId();
-    }
-
-    private void updateMemberPasswordWithVerificationCode(String verificationCode, Member member) {
-        member.updatePassword(verificationCode, passwordEncoder);
-        verificationService.deleteVerificationCode(verificationCode);
-    }
-}
